+++ src/components/GraphVisualisationWrapper.js
@@ -1,4 +1,4 @@
-import { useEffect } from 'react'
+import { useEffect, useRef } from 'react'
 import { connect } from 'redux-zero/react'
 import PropTypes from 'prop-types'
 import { useTranslation } from 'react-i18next'
@@ -7,11 +7,12 @@ import GraphVisualisation from './GraphVisualisation'
 // import jsonObjectProperties from '../assets/json/test-ontology-object-properties.json'
 import actions from '../store/actions'
 import setNodesIdsToDisplay from '../utils/setNodesIdsToDisplay'
-import { ALGO_TYPE_FULL } from '../constants/algorithms'
-import getAllTriplesPerNode from '../utils/getAllTriplesPerNode'
+import setGraphData from '../utils/setGraphData'
 import getGraphData from '../utils/getGraphData'
+import getNodeProperties from '../utils/getNodeProperties'
 import { SUB_CLASS_OF_ID, SUB_CLASS_OF_LABEL } from '../constants/graph'
 import GraphContextMenu from './GraphContextMenu'
+import loadGraphVersionFromServer from '../utils/versioning/loadGraphVersionFromServer'
 
 const GraphVisualisationWrapper = ({
   currentGraph,
@@ -19,14 +20,35 @@ const GraphVisualisationWrapper = ({
   setStoreState,
   showContextMenu,
   isBoundingBoxSelectable,
-  boundingBoxGeometry
+  boundingBoxGeometry,
+  addToObject,
+  selectedGraphVersion,
+  isOntologyUpdated
 }) => {
   const { t } = useTranslation()
+  const isInitialMountSelectedGraphVersion = useRef(true)
+  const isInitialMountOntologyUpdated = useRef(true)
+  const isInitialMountCurrentGraph = useRef(true)
 
   useEffect(async () => {
     const { classes, objectProperties } = await getGraphData({
       setStoreState,
       t
+    })
+
+    const nodesProperties = await getNodeProperties({
+      setStoreState,
+      t
+    })
+
+    setStoreState('nodesProperties', nodesProperties)
+
+    // TODO: Should become async when API call instead of localstorage
+    loadGraphVersionFromServer({
+      setStoreState,
+      addToObject,
+      classes,
+      objectProperties
     })
 
     // Set data from local file for debugging
@@ -39,40 +61,65 @@ const GraphVisualisationWrapper = ({
       rdfsLabel: SUB_CLASS_OF_LABEL
     }
 
-    setStoreState('classesFromApi', classes)
-    setStoreState('objectPropertiesFromApi', objectProperties)
-
-    const classesIds = Object.keys(classes)
-    const predicatesIds = Object.keys(objectProperties)
-
-    // in the background, parse classes to get triples per node
-    await getAllTriplesPerNode({
-      classesIds,
-      predicatesIds,
-      setStoreState,
-      classesFromApi: classes
-    })
-
-    setNodesIdsToDisplay({
-      type: ALGO_TYPE_FULL,
+    addToObject('graphVersions', 'original', {
       classesFromApi: classes,
       objectPropertiesFromApi: objectProperties,
-      setStoreState
+      classesFromApiBackup: classes,
+      objectPropertiesFromApiBackup: objectProperties,
+      deletedNodes: [],
+      addedNodes: [],
+      updatedNodes: []
+    })
+
+    setGraphData({
+      setStoreState,
     })
   }, [])
 
+  // Update nodes to display based on graph version except at component mount
   useEffect(() => {
-    // Update nodes to display based on selected graph
-    const {
-      type,
-      options
-    } = graphData[currentGraph]
+    if (isInitialMountSelectedGraphVersion.current) {
+      isInitialMountSelectedGraphVersion.current = false
+    } else {
+      setGraphData({
+        setStoreState
+      })
+    }
+  },
+  [
+    selectedGraphVersion
+  ])
 
-    setNodesIdsToDisplay({
-      type,
-      setStoreState,
-      options
-    })
+  // Update nodes to display based on graph version when ontology is updated except at component mount
+  useEffect(() => {
+    if (isInitialMountOntologyUpdated.current) {
+      isInitialMountOntologyUpdated.current = false
+    } else if (isOntologyUpdated) {
+      setGraphData({
+        setStoreState,
+      })
+      setStoreState('isOntologyUpdated', false)
+    }
+  },
+  [
+    isOntologyUpdated
+  ])
+
+  useEffect(() => {
+    if (isInitialMountCurrentGraph.current) {
+      isInitialMountCurrentGraph.current = false
+    } else {
+      const {
+        type,
+        options
+      } = graphData[currentGraph]
+
+      setNodesIdsToDisplay({
+        type,
+        setStoreState,
+        options
+      })
+    }
   }, [
     currentGraph
   ])
@@ -120,6 +167,9 @@ GraphVisualisationWrapper.propTypes = {
   showContextMenu: PropTypes.bool.isRequired,
   isBoundingBoxSelectable: PropTypes.bool.isRequired,
   boundingBoxGeometry: PropTypes.shape().isRequired,
+  addToObject: PropTypes.func.isRequired,
+  selectedGraphVersion: PropTypes.string.isRequired,
+  isOntologyUpdated: PropTypes.bool.isRequired,
 }
 
 const mapToProps = ({
@@ -128,14 +178,18 @@ const mapToProps = ({
   showContextMenu,
   contextMenuData,
   isBoundingBoxSelectable,
-  boundingBoxGeometry
+  boundingBoxGeometry,
+  selectedGraphVersion,
+  isOntologyUpdated
 }) => ({
   currentGraph,
   graphData,
   showContextMenu,
   contextMenuData,
   isBoundingBoxSelectable,
-  boundingBoxGeometry
+  boundingBoxGeometry,
+  selectedGraphVersion,
+  isOntologyUpdated
 })
 
 export default connect(