+++ src/components/NodesFilter.js
@@ -1,7 +1,6 @@
 /* eslint react/no-array-index-key:0 */
 import {
   useState,
-  useEffect
 } from 'react'
 import { connect } from 'redux-zero/react'
 import PropTypes from 'prop-types'
@@ -12,12 +11,11 @@ import { Button } from 'primereact/button'
 import actions from '../store/actions'
 import { SIDEBAR_VIEW_NODES_FILTER } from '../constants/views'
 import setFilteredNodes from '../utils/setFilteredNodes'
-import getNodesProperties from '../utils/getNodesProperties'
 
 const NodesFilter = ({
   setStoreState,
   nodesProperties,
-  updateGraphData
+  addToObject
 }) => {
   const { t } = useTranslation()
 
@@ -29,12 +27,6 @@ const NodesFilter = ({
   const checkEmptyRow = (filters) => filters.filter((filter) => filter.property === '' && filter.value === '').length > 0
 
   const [nodesFilters, setNodesFilters] = useState([defaultNodeFilter])
-
-  useEffect(() => {
-    if (nodesProperties.length === 0) {
-      getNodesProperties({ setStoreState })
-    }
-  }, [])
 
   return (
     <>
@@ -88,6 +80,8 @@ const NodesFilter = ({
                         id={selectId}
                         value={nodesFilters[index].property}
                         options={nodesProperties}
+                        optionValue="id"
+                        optionLabel="label"
                         onChange={(e) => {
                           const newFilter = {
                             ...nodesFilters[index],
@@ -152,7 +146,7 @@ const NodesFilter = ({
           onClick={() => setFilteredNodes({
             setStoreState,
             nodesFilters,
-            updateGraphData
+            addToObject
           })}
         />
       </div>
@@ -163,7 +157,7 @@ const NodesFilter = ({
 NodesFilter.propTypes = {
   setStoreState: PropTypes.func.isRequired,
   nodesProperties: PropTypes.arrayOf(PropTypes.string).isRequired,
-  updateGraphData: PropTypes.func.isRequired,
+  addToObject: PropTypes.func.isRequired,
 }
 
 const mapToProps = ({