+++ src/utils/setFilteredNodes.js
@@ -7,13 +7,13 @@ import store from '../store'
  * @param  {Object}   params
  * @param  {Array}    params.nodesFilters              Array of node filters {property [string], value [string]}
  * @param  {Function} params.setStoreState             setStoreState action
- * @param  {Function} params.updateGraphData           update graph data function
+ * @param  {Function} params.addToObject           Add to object action
  * @return
  */
 const setFilteredNodes = ({
   setStoreState,
   nodesFilters,
-  updateGraphData
+  addToObject
 }) => {
   const {
     lastGraphIndex,
@@ -33,7 +33,7 @@ const setFilteredNodes = ({
     }
   }
 
-  updateGraphData(newCurrentGraph, graphValue)
+  addToObject('graphData', newCurrentGraph, graphValue)
   setStoreState('currentGraph', newCurrentGraph)
   setStoreState('lastGraphIndex', newGraphIndex)
   setStoreState('sidebarView', SIDEBAR_VIEW_GRAPHS)