+++ CHANGELOG.md
@@ -18,6 +18,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.2.0]
 ### Added
+- ONTOLOGY-32 Added graph version save/upload functions, both to/from file and localStorage
+- ONTOLOGY-32 Added node create-update-delete-restore operations in selected graph version
+- ONTOLOGY-32 Added graph versioning tab and way to add/select version
 - ONTOLOGY-21 Filter nodes by properties
 - ONTOLOGY-22 Filter edges by properties
 - ONTOLOGY-51 Added JSDoc build command
@@ -33,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 ### Fixed
+- ONTOLOGY-32 Added useRef constants in GraphVisualisationWrapper to avoid refreshing the graph multiple times at startup
 - ONTOLOGY-47 Removed highlight from free-text search selected node after sidebar section update
 
 ### Security